tracing/pyproject.toml
[project]
name = "tracing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.1.0",
    "sdk>=1.0.0",
]
